Updated pricing structure
pricing is now structured in a more OOP way to better incorporate future resource implementation.
