+++ samcli/commands/check/resources/unique_pricing_info.py
@@ -3,13 +3,13 @@ A super class for all sub pricing classes. When a new resource is added and pric
 is implemebted for it, all common methods between resoruces will be implemented here
 """
 
+from abc import abstractmethod
+
 
 class UniquePricingInfo:
-    """
-    Note: Currently, pricing is only implemented for lambda functions. Until pricing
-    is implemeted for more resources, this class will not be utilized., since everything
-    implemeted for lambda functions is currently unique to lambda functions.
-    """
-
     def __init__(self):
+        pass
+
+    @abstractmethod
+    def ask_questions(self):
         pass
