+++ tests/unit/commands/check/test_command_context.py
@@ -13,7 +13,7 @@ from samcli.commands.check.lib.command_context import CheckContext, _parse_templ
 
 class TestCommandContext(TestCase):
     @patch("samcli.commands.check.lib.command_context.CheckResults")
-    @patch("samcli.commands.check.lib.command_context.PricingCalculations")
+    @patch("samcli.commands.check.lib.command_context.LambdaFunctionPricingCalculations")
     @patch("samcli.commands.check.lib.command_context.BottleNeckCalculations")
     @patch("samcli.commands.check.lib.command_context._parse_template")
     @patch("samcli.commands.check.lib.command_context.BottleNecks")
@@ -22,7 +22,7 @@ class TestCommandContext(TestCase):
         patch_bottle_neck,
         patch_parse_template,
         patch_bottle_neck_calculations,
-        patch_pricing_calculations,
+        patch_lambda_pricing_calculations,
         patch_print,
     ):
         region = Mock()
@@ -42,8 +42,8 @@ class TestCommandContext(TestCase):
         patch_bottle_neck_calculations.run_bottle_neck_calculations = Mock()
         patch_print.print_bottle_neck_results = Mock()
 
-        patch_pricing_calculations.run_calculations = Mock()
-        patch_pricing_calculations.get_lambda_pricing_results = Mock()
+        patch_lambda_pricing_calculations.run_calculations = Mock()
+        patch_lambda_pricing_calculations.lambda_pricing_results = Mock()
 
         patch_parse_template.return_value = graph_mock
