+++ tests/unit/commands/check/test_pricing_calculations.py
@@ -1,37 +1,32 @@
 from unittest import TestCase
 from unittest.mock import Mock, patch
-from samcli.commands.check.pricing_calculations import PricingCalculations, _convert_usage_type
+from samcli.commands.check.lambda_function_pricing_calculations import (
+    LambdaFunctionPricingCalculations,
+    _convert_usage_type,
+)
 
 
-class TestPricingCalculations(TestCase):
-    def test_get_lambda_pricing_results(self):
-        self_mock = Mock()
-        self_mock.lambda_pricing_results = Mock()
-
-        result = PricingCalculations.get_lambda_pricing_results(self_mock)
-
-        self.assertEqual(result, self_mock.lambda_pricing_results)
-
+class TestLambdaFunctionPricingCalculations(TestCase):
     def test_run_calculations(self):
         self_mock = Mock()
         self_mock.get_charge_and_request_amounts = Mock()
-        self_mock.determine_lambda_cost = Mock()
+        self_mock.determine_cost = Mock()
 
-        PricingCalculations.run_calculations(self_mock)
+        LambdaFunctionPricingCalculations.run_calculations(self_mock)
 
         self_mock.get_charge_and_request_amounts.assert_called_once()
-        self_mock.determine_lambda_cost.assert_called_once()
+        self_mock.determine_cost.assert_called_once()
 
     def test_get_charge_and_request_amounts(self):
         self_mock = Mock()
         self_mock.get_lambda_charge_and_request_amounts = Mock()
 
-        PricingCalculations.get_charge_and_request_amounts(self_mock)
+        LambdaFunctionPricingCalculations.get_charge_and_request_amounts(self_mock)
 
         self_mock.get_lambda_charge_and_request_amounts.assert_called_once()
 
-    @patch("samcli.commands.check.pricing_calculations.ast")
-    @patch("samcli.commands.check.pricing_calculations.urllib")
+    @patch("samcli.commands.check.lambda_function_pricing_calculations.ast")
+    @patch("samcli.commands.check.lambda_function_pricing_calculations.urllib")
     def test_get_aws_lambda_pricing_info(self, patch_request, patch_ast):
 
         file_return_mock = Mock()
@@ -39,7 +34,7 @@ class TestPricingCalculations(TestCase):
 
         patch_ast.literal_eval.return_value = Mock()
 
-        result = PricingCalculations.get_aws_lambda_pricing_info(Mock(), "")
+        result = LambdaFunctionPricingCalculations.get_aws_lambda_pricing_info(Mock(), "")
 
         self.assertEqual(result, patch_ast.literal_eval.return_value)
 
@@ -58,12 +53,12 @@ class TestPricingCalculations(TestCase):
 
         self_mock.graph.unique_pricing_info = {"LambdaFunction": template_lambda_pricing_info_mock}
 
-        PricingCalculations.determine_lambda_cost(self_mock)
+        LambdaFunctionPricingCalculations.determine_lambda_cost(self_mock)
 
         self.assertEqual(381.38, self_mock.lambda_pricing_results)
 
-    @patch("samcli.commands.check.pricing_calculations._convert_usage_type")
-    @patch("samcli.commands.check.pricing_calculations.get_session")
+    @patch("samcli.commands.check.lambda_function_pricing_calculations._convert_usage_type")
+    @patch("samcli.commands.check.lambda_function_pricing_calculations.get_session")
     def test_get_lambda_charge_and_request_amounts(self, patch_session, patch_convert):
         self_mock = Mock()
         self_mock.get_region_prefix = Mock()
@@ -80,19 +75,19 @@ class TestPricingCalculations(TestCase):
 
         patch_session.get_config_variable.return_value = ""
 
-        self_mock.get_aws_lambda_pricing_info = Mock()
-        self_mock.get_aws_lambda_pricing_info.return_value = {"products": products, "terms": terms_mock}
+        self_mock.get_aws_pricing_info = Mock()
+        self_mock.get_aws_pricing_info.return_value = {"products": products, "terms": terms_mock}
 
         patch_convert.return_value = ["", "Lambda-GB-Second"]
 
-        PricingCalculations.get_lambda_charge_and_request_amounts(self_mock)
+        LambdaFunctionPricingCalculations.get_lambda_charge_and_request_amounts(self_mock)
 
         self_mock.get_pricing_or_request_value.assert_any_call(product_1, terms_mock, "global-request")
         self_mock.get_pricing_or_request_value.assert_any_call(product_2, terms_mock, "global-lambda")
         self_mock.get_pricing_or_request_value.assert_any_call(product_3, terms_mock, "lambda")
 
         patch_convert.return_value = ["Request", ""]
-        PricingCalculations.get_lambda_charge_and_request_amounts(self_mock)
+        LambdaFunctionPricingCalculations.get_lambda_charge_and_request_amounts(self_mock)
         self_mock.get_pricing_or_request_value.assert_any_call(product_4, terms_mock, "request")
 
     def test_get_pricing_or_request_value(self):
@@ -108,19 +103,19 @@ class TestPricingCalculations(TestCase):
         terms = {"OnDemand": {sku: temp_key}}
 
         # testing global-request
-        PricingCalculations.get_pricing_or_request_value(self_mock, product, terms, "global-request")
+        LambdaFunctionPricingCalculations.get_pricing_or_request_value(self_mock, product, terms, "global-request")
         self.assertEqual(8.15, self_mock.max_num_of_free_requests)
 
         # testing global-lambda
-        PricingCalculations.get_pricing_or_request_value(self_mock, product, terms, "global-lambda")
+        LambdaFunctionPricingCalculations.get_pricing_or_request_value(self_mock, product, terms, "global-lambda")
         self.assertEqual(8.15, self_mock.max_free_GBs)
 
         # testing lambda
-        PricingCalculations.get_pricing_or_request_value(self_mock, product, terms, "lambda")
+        LambdaFunctionPricingCalculations.get_pricing_or_request_value(self_mock, product, terms, "lambda")
         self.assertEqual(115, self_mock.monthly_compute_charge)
 
         # testing request
-        PricingCalculations.get_pricing_or_request_value(self_mock, product, terms, "request")
+        LambdaFunctionPricingCalculations.get_pricing_or_request_value(self_mock, product, terms, "request")
         self.assertEqual(115, self_mock.monthly_request_charge)
 
     def test_convert_usage_type(self):
